CryptoTrading/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
binance-futures-connector==3.2.0
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
dateparser==1.1.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
frozenlist==1.3.1
gevent==21.12.0
greenlet==1.1.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.15.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.2
packaging==21.3
pandas==1.4.4
pandas-ta==0.3.14b0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.7
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.3.2
rel==0.4.7
requests==2.28.1
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.4.0
websockets==10.3
widgetsnbextension==4.0.2
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.4.0
